% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{SapiClient-class}
\alias{SapiClient}
\alias{SapiClient-class}
\title{Client for working with Keboola Connection Storage API.}
\description{
Client for working with Keboola Connection Storage API.
}

